+++ resources/views/admin/pages/proposal/index.blade.php
@@ -79,10 +79,10 @@
             },
             {
                 data: 'status',
-                render: function(data,type,row){
-                    if(data == 1){
+                render: function (data, type, row) {
+                    if (data == 1) {
                         return '<div class="badge bg-success">Aktif</div>';
-                    }else{
+                    } else {
                         return '<div class="badge bg-danger">Tidak Aktif</div>';
                     }
                 }
@@ -115,21 +115,40 @@
     });
 
     hapus = (id) => {
-        $.ajax({
-            // headers: {
-            //     'X-CSRF-TOKEN': $('meta[name="csrf_token"]').attr('content'),
-            // },
-            type: "delete",
-            url: "/api/proposal/" + id+"/delete",
-            success: function (data) {
+        Swal.fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, delete it!"
+        }).then((result) => {
+            if (result.isConfirmed) {
+                $.ajax({
+                    type: "delete",
+                    url: "/api/proposal/" + id + "/delete",
+                    success: function (data) {
+                        table.ajax.reload()
+                        Swal.fire({
+                            title: "Success!",
+                            text: "Data Berhasil Di hapus",
+                            icon: "success"
+                        });
+                    },
+                    error: function (jqxhr, textStatus, error) {
+                        var err = textStatus + ", " + error;
+                        Swal.fire({
+                            title: "Failed!",
+                            text: "Data Gagal Di hapus",
+                            icon: "error"
+                        });
+                    }
+                })
 
-                table.ajax.reload()
-            },
-            error: function (jqxhr, textStatus, error) {
-                var err = textStatus + ", " + error;
-                console.log("Request Failed: " + err);
+                //
             }
-        })
+        });
     }
     // search engine
     // $("#search").keyup(function () {
